+++ src/controllers/patient/preConsultation/ProvisionalPreConsultationController.js
@@ -4,14 +4,11 @@ import createPatientPainMapHandler from "../../../handlers/painMap/createPatient
 
 const createPreConsultationController = async (req, res) => {
   try {
-    const validate = validateDuplicatePainArea(req.body.painRecordsToCreate[0]);
-
-    if(validate===false)throw new Error("Area de dolor duplicada")
-
     const preConsultation = await createPreConsultationHandler(req.body);
     const vitalSigns = await createVitalSignsHandler(
       req.body.vitalSignsToCreate
     ); //It receives an array of vital signs
+
     const physicalExamination = await createPatientPainMapHandler(
       req.body.painRecordsToCreate[0]
     );
@@ -20,19 +17,10 @@ const createPreConsultationController = async (req, res) => {
       .status(201)
       .json({ preConsultation, vitalSigns, physicalExamination });
   } catch (error) {
-    return res.status(500).json({"Hubo un error durante el proceso de creación: ": error.message})
+    return res.status(500).json({
+      "Hubo un error durante el proceso de creación: ": error.message,
+    });
   }
 };
 
 export default createPreConsultationController;
-
-const validateDuplicatePainArea = (painRecordsToCreate) => {
-  const seenPainAreas = new Set();
-  for (const obj of painRecordsToCreate.painAreas) {
-      if (seenPainAreas.has(obj.painArea)) {
-      return false;
-    }
-    seenPainAreas.add(obj.painArea);
-  }
-  return true;
-};
