+++ src/handlers/painMap/createPatientPainMapHandler.js
@@ -2,13 +2,20 @@ import SegimedAPIError from "../../error/SegimedAPIError.js";
 import contextService from "request-context";
 import moment from "moment-timezone";
 import { PatientPainMap } from "../../databaseConfig.js";
+import validateDuplicatePainArea from "../../validations/validateDuplicatePainArea.js";
 
-const createPatientPainMapHandler = async (body) => {
-  const patientPainMapping = mapPainRecord(body);
+const createPatientPainMapHandler = async (painRecord) => {
+  const patientPainMapping = mapPainRecord(painRecord);
+
   try {
+    // Validate if the pain area is duplicated
+    const validate = validateDuplicatePainArea(painRecord);
+    if (validate === false) throw new Error("Area de dolor duplicada");
+    // Create the new pain record
     const newPainRecords = await PatientPainMap.create(patientPainMapping);
     return newPainRecords;
   } catch (error) {
+    console.log(error);
     throw new SegimedAPIError(
       "Hubo un error durante el proceso de creación.",
       500
